Move Goal and ProofState to their own packages.
