+++ kore/test/Test/Kore/AllPath.hs
@@ -15,19 +15,19 @@ import qualified Data.Sequence as Seq
 import           GHC.Stack
                  ( HasCallStack )
 
-import qualified Kore.Goal as Goal
 import qualified Kore.Internal.MultiOr as MultiOr
 import           Kore.Logger
                  ( LogMessage (..), WithLog (..) )
 import           Kore.Profiler.Data
                  ( Configuration (..), MonadProfiler (..) )
-import qualified Kore.ProofState as Goal
 import           Kore.Step.Simplification.Data
                  ( MonadSimplify (..) )
 import qualified Kore.Step.Strategy as Strategy
 import           Kore.Step.Transition
                  ( runTransitionT )
 import qualified Kore.Step.Transition as Transition
+import qualified Kore.Strategies.Goal as Goal
+import qualified Kore.Strategies.ProofState as Goal
 import           SMT
                  ( MonadSMT (..) )
